💀

+++ src/pages/api/comment/[id].js
@@ -10,6 +10,7 @@ export default async function handler(req, res) {
 
   switch (method) {
     case "GET":
+      // if (session) {
       const comment = await prisma.comment.findUnique({
         where: {
           id: Number.parseInt(commentId),
@@ -24,6 +25,9 @@ export default async function handler(req, res) {
       } else {
         res.status(200).json(comment);
       }
+      // } else {
+      res.status(401).end("Unauthorized");
+      // }
       break;
 
     case "PUT":
@@ -60,12 +64,16 @@ export default async function handler(req, res) {
       break;
 
     case "DELETE":
-      const deletedComment = await prisma.comment.delete({
-        where: {
-          id: Number.parseInt(query.id),
-        },
-      });
-      res.status(200).json(deletedComment);
+      if (session) {
+        const deletedComment = await prisma.comment.delete({
+          where: {
+            id: Number.parseInt(query.id),
+          },
+        });
+        res.status(200).json(deletedComment);
+      } else {
+        res.status(401).end("Unauthorized");
+      }
       break;
 
     default:
